// app/(dashboard)/certificates/electrical-installation-condition-report/[id]/update/schedule-of-inspections/section-5/schema.ts
import { z } from "zod";

export const Schema = z.object({
  item_5_1: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_2: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_3: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_4: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_5: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_6: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_7: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_8: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_9: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_10: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_11: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_12: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_13: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_14A: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_14B: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_15: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_16: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_17: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_18: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_19: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_20: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_21: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_22: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_23: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
  item_5_24: z.enum(["c1", "c2", "c3", "na", "lim", "r"] as const),
});